lwio: CID# 11707, NO_EFFECT.
git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@52725 9fc1c11e-11b2-4e14-94d8-6c5e05775969

+++ lwio/utils/time.c
@@ -82,7 +82,7 @@ LwioGetSystemTimeString(
     NTSTATUS status = STATUS_SUCCESS;
     struct timespec ts = {0};
     struct tm tmbuf = {0};
-    size_t sWritten = 0;
+    LONG sWritten = 0;
 
     status = LwioGetSystemTime(&ts);
     if (status)
